# Conflicts:
#	train.py

+++ data/left_contiguous_csr.py
@@ -1,3 +1,5 @@
+# Branch keshav
+
 import numpy as np
 
 
@@ -18,7 +20,7 @@ class LeftContiguousCSR:
     @staticmethod
     def load(filepath):
         npzfile = np.load(filepath)
-        return LeftContiguousCSR(npzfile['indptr'], npzfile['degrees'], npzfile['data'])
+        return LeftContiguousCSR(npzfile['indptr'].astype(np.int64), npzfile['degrees'].astype(np.int64), npzfile['data'].astype(np.int64))
 
 # uniform batch sampling can be done as follows
